Various changes in TVM, github builds and tests
* Bugfixes in TVM and node

* Upgrade to C++17

* Improve GitHub builds

* Fix existing tests and partially integrate them into builds

---------

+++ tdfec/td/fec/raptorq/Rfc.h
@@ -61,7 +61,7 @@ class Rfc {
     template <class F>
     void encoding_row_for_each(EncodingRow t, F &&f) const {
       f(t.b);
-      for (uint16 j = 1; j < t.d; ++j) {
+      for (uint32 j = 1; j < t.d; ++j) {
         t.b = (t.b + t.a) % W;
         f(t.b);
       }
